Only send mqtt message when meeting status changes
Prevously the zoom-detector script was sending a message to the queue
each time through the loop. This is obviously not necessary!
Made the output less noisy.
Catch exceptions around detecting processes. SOmetimes when iterating
through the list of processes, the current process has ended by the time
the code to look at the name executes and causes an exception. I wasn't
able to figure out how to catch the specific exception that was being
thrown, so I'm catching all types. Not the greatest code, but it should
guard against crashes.